+++ middleware2/src/main/java/expense/data/ExpenseTypeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ExpenseTypeRepository extends JpaRepository<ExpenseType,Long> {
+public interface ExpenseTypeRepository extends JpaRepository<ExpenseType,Integer> {
 }
